splitting code up into modules and added effects if helper
