+++ README.md
@@ -411,6 +411,32 @@ withEffects(app)({
 }).foo()
 ```
 
+### `effectsIf`
+
+```js
+EffectConditional = [boolean, EffectTuple]
+effectsIf = function(EffectConditional[]): EffectTuple[]
+```
+
+Convert an array of `[boolean, EffectTuple]`s into a new array of effects where the boolean evaluates to true. This provides compact syntatic sugar for conditionally firing effects.
+
+Example:
+
+```js
+import { withEffects, effectsIf, action } from "hyperapp-effects"
+
+withEffects(app)({
+  actions: {
+    foo: () => ({ running }) => effectsIf([
+      [true, action("always")],
+      [false, action("never")],
+      [running, action("ifRunning")],
+      [!running, action("ifNotRunning")]
+    ])
+  }
+}).foo()
+```
+
 ## License
 
 Hyperapp Effects is MIT licensed. See [LICENSE](LICENSE.md).
