+++ src/effectTypes.js
@@ -1,0 +1,10 @@
+export var ACTION = "action"
+export var FRAME = "frame"
+export var DELAY = "delay"
+export var TIME = "time"
+export var LOG = "log"
+export var HTTP = "http"
+export var EVENT = "event"
+export var KEY_DOWN = "keydown"
+export var KEY_UP = "keyup"
+export var RANDOM = "random"
